pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-rag-chatbot"
version = "0.1.0"
description = "A Multi agent RAG-based chatbot service"
requires-python = ">=3.9"

dependencies = [
    "fastapi[standard]==0.122.0",
    "langchain==0.3.27",
    "langchain-chroma==0.2.6",
    "langchain-community==0.3.31",
    "langchain-openai==0.3.35",
    "langfuse>=2.0.0",
    "python-dotenv==1.0.0",
    "tiktoken>=0.5.0",
    "slowapi>=0.1.9"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]
